# server/openapi.yaml
...
openapi: 3.1.0
info:
  title: Fake Stack Overflow API
  version: 1.0.0
  description: OpenAPI 3.1 specification for the Fake Stack Overflow backend server.
servers:
  - url: /api
    description: The API server endpoint
paths:
...
  /answer/addAnswer:
    post:
      tags:
        - Answers
      summary: Add a new answer to a specific question.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAnswerBody'
      responses:
        '200':
          description: Answer successfully added. Returns the database answer object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseAnswer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /answer/delete/{aid}:
    delete:
      tags:
        - Answers
      summary: Delete an answer created by the requesting user.
      parameters:
        - in: path
          name: aid
          schema:
            type: string
            format: object-id
          required: true
          description: The unique ID of the answer to delete.
        - in: query
          name: username
          schema:
            type: string
          required: true
          description: Username of the user attempting to delete the answer.
      responses:
        '200':
          description: Answer successfully deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAnswerResult'
        '403':
          description: The user is not allowed to delete this answer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorBody'
        '404':
          description: Answer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleErrorBody'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  # --------------------
  # TAG CONTROLLER
  # --------------------
...
components:
  schemas:
...
    AnswerBody:
      type: object
      required:
        - text
        - ansBy
        - ansDateTime
      properties:
        text:
          type: string
          minLength: 1
        ansBy:
          type: string
          minLength: 1
        ansDateTime:
          type: string
          format: date-time
    AddAnswerBody:
      type: object
      required:
        - qid
        - ans
      properties:
        qid:
          type: string
          format: object-id
          description: The ID of the question the answer belongs to.
        ans:
          $ref: '#/components/schemas/AnswerBody'
    DatabaseAnswer:
      type: object
      description: An answer object as stored in MongoDB.
      required:
        - _id
        - text
        - ansBy
        - ansDateTime
        - comments
      properties:
        _id:
          type: string
          format: object-id
        text:
          type: string
        ansBy:
          type: string
        ansDateTime:
          type: string
          format: date-time
        comments:
          type: array
          items:
            type: string
            format: object-id

    # --------------------
    # TAG SCHEMAS
    # --------------------
...
    DeleteAnswerResult:
      type: object
      properties:
        success:
          type: boolean
          example: true
        deletedAnswerId:
          type: string
          format: object-id

    # --------------------
    # USER SCHEMAS
    # --------------------
...
    SimpleErrorBody: # MODIFICATION: New schema for simple JSON errors like 403, 404
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: A detailed error message.
  responses:
...
    ErrorResponse:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: The error message returned by the server.
    ValidationError:
      description: Bad Request / Validation Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: The error message from express-openapi-validator.
              errors:
                type: array
                description: Detailed list of validation errors.
...